+++ corehq/motech/dhis2/const.py
@@ -21,3 +21,5 @@ DHIS2_EVENT_STATUSES = (
 )
 
 LOCATION_DHIS_ID = 'dhis_id'
+
+DHIS2_API_VERSION = '26'
