+++ corehq/motech/openmrs/openmrs_config.py
@@ -2,41 +2,20 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 from corehq.motech.openmrs.finders import PatientFinder
+from corehq.motech.value_source import ValueSource
 from dimagi.ext.couchdbkit import (
     DictProperty,
     DocumentSchema,
     ListProperty,
     SchemaDictProperty,
-    SchemaListProperty,
     SchemaProperty,
     StringProperty,
 )
 
 
-def recurse_subclasses(cls):
-    return (
-        cls.__subclasses__() +
-        [subsub for sub in cls.__subclasses__() for subsub in recurse_subclasses(sub)]
-    )
-
-
 class IdMatcher(DocumentSchema):
     case_property = StringProperty()
     identifier_type_id = StringProperty()
-
-
-class ValueSource(DocumentSchema):
-    @classmethod
-    def wrap(cls, data):
-        if cls is ValueSource:
-            return {
-                sub._doc_type: sub for sub in recurse_subclasses(cls)
-            }[data['doc_type']].wrap(data)
-        else:
-            return super(ValueSource, cls).wrap(data)
-
-    def get_value(self, case_trigger_info):
-        raise NotImplementedError()
 
 
 class CaseProperty(ValueSource):
