+++ corehq/motech/value_source.py
@@ -1,0 +1,25 @@
+from __future__ import absolute_import
+
+from __future__ import unicode_literals
+from couchdbkit.ext.django.schema import DocumentSchema
+
+
+def recurse_subclasses(cls):
+    return (
+        cls.__subclasses__() +
+        [subsub for sub in cls.__subclasses__() for subsub in recurse_subclasses(sub)]
+    )
+
+
+class ValueSource(DocumentSchema):
+    @classmethod
+    def wrap(cls, data):
+        if cls is ValueSource:
+            return {
+                sub._doc_type: sub for sub in recurse_subclasses(cls)
+            }[data['doc_type']].wrap(data)
+        else:
+            return super(ValueSource, cls).wrap(data)
+
+    def get_value(self, case_trigger_info):
+        raise NotImplementedError()
